modified user service authorization logic

+++ src/main/java/pl/pmar/blogplatform/service/UserService.java
@@ -1,6 +1,8 @@
 package pl.pmar.blogplatform.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pl.pmar.blogplatform.model.entity.Post;
@@ -15,23 +17,40 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final SecurityContextService contextService;
 
     @Autowired
-    public UserService(UserRepository userRepository, PostRepository postRepository) {
+    public UserService(
+            UserRepository userRepository,
+            PostRepository postRepository,
+            SecurityContextService contextService
+    ) {
         this.userRepository = userRepository;
         this.postRepository = postRepository;
+        this.contextService = contextService;
     }
 
     public ResponseEntity<List<User>> getAllUsers() {
-        List<User> users = userRepository.findAll();
-        return ResponseEntity.ok(users);
+        if (contextService.isUserAdmin()) {
+            List<User> users = userRepository.findAll();
+            return ResponseEntity.ok(users);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
     }
 
     public ResponseEntity<User> getUserById(Integer id) {
-        Optional<User> user = userRepository.findById(id);
-        return user
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        Optional<User> userOptional = userRepository.findById(id);
+        if (userOptional.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        User user = userOptional.get();
+        if (contextService.isUserAuthorized(user.getId())) {
+            return ResponseEntity.ok(user);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
     }
 
@@ -39,12 +58,15 @@ public class UserService {
         Optional<User> user = userRepository.findById(id);
         if (user.isEmpty()) {
             return ResponseEntity.notFound().build();
-        } else {
-            List<Post> userPosts = postRepository.findAllByAuthorId(id);
-            postRepository.deleteAll(userPosts);
-            userRepository.deleteById(id);
-            return ResponseEntity.noContent().build();
         }
+        if (!contextService.isUserAuthorized(id)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        List<Post> userPosts = postRepository.findAllByAuthorId(id);
+        postRepository.deleteAll(userPosts);
+        userRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
 
     }
 
@@ -52,12 +74,14 @@ public class UserService {
         Optional<User> userOptional = userRepository.findById(updatedUser.getId());
         if (userOptional.isEmpty()) {
             return ResponseEntity.notFound().build();
-        } else {
-            User user = userOptional.get();
-            user.setUsername(updatedUser.getUsername());
-            user.setEmail(updatedUser.getEmail());
-
-            return ResponseEntity.ok(userRepository.save(user));
         }
+        User user = userOptional.get();
+        if (!contextService.isUserAuthorized(user.getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        user.setUsername(updatedUser.getUsername());
+        user.setEmail(updatedUser.getEmail());
+
+        return ResponseEntity.ok(userRepository.save(user));
     }
 }
